requirements.txt
google-generativeai
python-dotenv
google-cloud-aiplatform
requests
google-adk==0.5.0
